+++ RELEASE_NOTES.md
@@ -1,5 +1,17 @@
 Release Notes
 =============
+## 1.2.0-beta3
+* SQL Azure: Connection string owner now has the correct path.
+* SQL Azure: New PasswordParameter returns the name of the Password parameter.
+* Functions: Ability to override the auto-generated storage account name.
+* Functions: Ability to add multiple ARM Expressions as settings.
+* Functions: Ability to add a Resource Name as a setting.
+* WebApp: Ability to add multiple ARM Expressions as settings.
+* Storage Account: WebsitePrimaryEndpoint is now a generated ARM expression.
+* Add branding.
+
+* Internal: ARM Expression refernces now add the Resource Id as the Owner.
+
 ## 1.2.0-beta2
 * Log Analytics: Initial release.
 * Static Web Apps: Initial release.
@@ -19,7 +31,7 @@ Release Notes
 * Event Hubs: Remove redundant kafka flag (minor breaking change).
 
 * Internal changes to better capture full resource IDs.
-* Add branding.
+* Storage Account: WebsitePrimaryEndpoint depends on storage account name instead of being hardcoded.
 
 ## 1.2.0-beta1
 * Bastion Hosts: Create bastion hosts for accessing resources on a virtual network